Move email template overrides functions to EmailImprovements class
